i#2266 memlayout: Add callstack frames to layout dump
Updates DR to 863a461c for dr_annotation_pass_pc().  Uses it to
acquire the PC for the current thread for the memory layout dump.

Adds callstack frames that overlap with the dumped stack to the layout
dump, with a pc, frame pointer, and function name for each.

Adds a function to tests/memlayout to test an additional frame.

Issue: #2266