i#2317: Fix shadow_check_range() crash
Fixes a bug where shadow_check_range() bypasses the shadow layer that
handles lazily-allocated shadow mappings and thus can crash.

Adds a test that reproduces the bug.

Adds a new DRMEMORY_ANNOTATE_MAKE_UNADDRESSABLE annotation to make it
easier to write the test.

Updates the allocator to use uint instead of ushort for its header's
request_diff field, as ushort is too small for a large allocation like
the one used in the test.  Since that affects the layout, goes ahead
and updates the prev_size_shr field to uint as well.

Fixes #2317