i#2356: Add md5 of initial ELF bytes to drsymcache
Previously, for Linux drsymcache only checked the library file size to
determine whether the library had changed.  This led to bogus values
and subsequent false positives and worse problems.  Here we add an md5
computed over the first 1024 bytes of the mapped-in library as a
checksum.  It experimentally is enough to distinguish similar libc
versions (which have identical file sizes and top-level ELF headers).

Tested on Debian libc.so.6 2.31-5 vs 2.31-9.

Fixes #2356