i#2363 vsyscall: Remove umbra gap without vsyscall
Checks for the vsyscall and if it is not present, removes the
0x7ff'ff4-0x7ff'ff8 gap from Umbra's mappings.  This then allows the
kernel or the app to map into the gap region, avoiding Umbra failures
in such circumstances.

Generalizes the Umbra region merging to also merge non-pre-defined
regions to handle the test added for #1712.

Tested by running tests/mmap in a loop, which found the original
problem.

The problem of vsyscall being present and the kernel mapping vdso into
the gap is left unsolved here.

Issue: #2363