Update mangled operator names for x64 (#2369)
The size_t type's mangled letter is m for 64-bit in the Itanium ABI
style.  We update the operator new mangled names for 64-bit to s/j/m/
here.  These are only used for distinguishing nothrow and placement,
not for locating operators, which may explain why there was no noticed
problem in the past.