i#2080: Separate default and user-set blocklists
Moves the blocklist filled with default Windows paths to a new option
-lib_blocklist_default.  This lets the user add to the blocklist
without having to figure out what the default was and re-list it.

Updates the documentation.

Updates the existing blocklist test to test the new option behavior.
Relaxes the existing blocklist Windows test output so that it can be
removed from the ignore list.
Adds a new cross-platform test of the base blocklist feature.

Fixes #2080