i#2421: Handle wild accesses in Umbra
Adds decoding of cache pc's to identify the AND;ADD;SHIFT sequence
used for inlined x86-64 shadow translation in order to add a new
shadow region for a wild application access, preventing passing a
shadow memory access crash to the app.

Adds a new test of an app that accesses unmapped memory outside of the
typical ranges and fails if the fault address does not match what it
accessed.

This also fixes the malloc and related tests on Ubuntu18+ where 0x100
is not already added and hits this wild address path.

Issue: #2421, #2417
Fixes #2421