i#2498 init failures: Ignore privlib init failures (#2515)
Updates DR to 51b152fa2 extends the private loader bcrypt.dll entry
failure being ignored to ignore all private library entry failures. We
have had multiple cases of initializers failing without fatal
consequences, and given that we do not have resources to track down
every detail it is best to note the failure and move on. We have seen
Dr. Memory in particular fail on newer Windows versions and in some
cases ignoring that failure lets the tool continue and succeed, as only
a small piece of the private library is needed (generally pulled in as a
chain of dependences from dbghelp.dll which is used for symbol
information). The private loader on Windows is considered best-effort at
this point.

Issue: #2498