i#2478: Fix Compilation error on linux mint 21.1  (#2546)
The code changes involve updating the handling of the arg variable,
which is used in the handle_semctl function to represent the semun union
argument passed to the semctl system call. The main change is to
dereference a pointer to the arg variable instead of using a direct
structure assignment.

Fixes #2478