Update DR to fix Windows build warning (#2570)
Updates DR to 92be9e40f to pull in the fix for a drreg compiler warning
which is breaking the build,
along with the drmgr fix for the umbra_test_allscales failure.

Updates drmemory usage of several DR events to use drmgr instead to
comply with drmgr changes.

Avoids some double-definitions of macros.

Issue DynamoRIO/dynamorio#7739